# Optimize til::rle to do replacing a little faster