@flyingcat 2024-10-17
@lhecker Sorry for the late reply. I am going to close this PR, as I don't have much free time now and it's a pain to compile this project on a slow computer. After all, this is just a minor improvement of little value.

Also, I'm really not used to this inefficient way of communication. When I work, I communicate face to face or through instant messaging software.