@lhecker 2024-10-17
I apologize for the inconvenience, but I'm very grateful for the ideas you've proposed to us! I'll try to put them to good use. 🙂

I believe the reason it compiles so slowly is because that header is currently included everywhere. If we limited it to only those parts of the project that needed it, it would compile in seconds instead. I'll consider making that change in the future.