@flyingcat 2024-10-17
Yeah, part of it, but mostly because my computer is really old and slow. Before rich enough to get a 4080s GPU and 4k144hz monitor, I have no motivation to upgrade it. 

Any way, thanks for your reviews and reply.